Revert "feat(payment): PAYPAL-2843 captured device info for Braintree Apple Pay (#2131)"
This reverts commit 05fbd6ef754fd163b403cc958460e7ae4ec6031e.
