+++ packages/apple-pay-integration/jest.config.js
@@ -1,12 +1,12 @@
 module.exports = {
-    displayName: 'apple-pay',
-    preset: '../../jest.preset.js',
+    displayName: "apple-pay",
+    preset: "../../jest.preset.js",
     globals: {
-        'ts-jest': {
-            tsconfig: '<rootDir>/tsconfig.spec.json',
+        "ts-jest": {
+            tsconfig: "<rootDir>/tsconfig.spec.json",
             diagnostics: false,
         },
     },
-    setupFilesAfterEnv: ['../../jest-setup.js'],
-    coverageDirectory: '../../coverage/packages/apple-pay-integration',
+    setupFilesAfterEnv: ["../../jest-setup.js"],
+    coverageDirectory: "../../coverage/packages/apple-pay-integration",
 };
