+++ packages/apple-pay-integration/src/apple-pay-payment-strategy.ts
@@ -1,6 +1,5 @@
 import { RequestSender } from '@bigcommerce/request-sender';
 
-import { BraintreeIntegrationService } from '@bigcommerce/checkout-sdk/braintree-utils';
 import {
     InvalidArgumentError,
     NotInitializedError,
@@ -18,7 +17,6 @@ import {
     PaymentStrategy,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
 
-import { ApplePayGatewayType } from './apple-pay';
 import { WithApplePayPaymentInitializeOptions } from './apple-pay-payment-initialize-options';
 import ApplePaySessionFactory from './apple-pay-session-factory';
 
@@ -45,7 +43,6 @@ export default class ApplePayPaymentStrategy implements PaymentStrategy {
         private _requestSender: RequestSender,
         private _paymentIntegrationService: PaymentIntegrationService,
         private _sessionFactory: ApplePaySessionFactory,
-        private _braintreeIntegrationService: BraintreeIntegrationService,
     ) {}
 
     async initialize(
@@ -64,14 +61,6 @@ export default class ApplePayPaymentStrategy implements PaymentStrategy {
         this._storeCreditLabel = options.applepay?.storeCreditLabel || DefaultLabels.StoreCredit;
 
         await this._paymentIntegrationService.loadPaymentMethod(methodId);
-
-        const state = await this._paymentIntegrationService.loadPaymentMethod(methodId);
-
-        const paymentMethod: PaymentMethod = state.getPaymentMethodOrThrow(methodId);
-
-        if (paymentMethod.initializationData?.gateway === ApplePayGatewayType.BRAINTREE) {
-            await this._initializeBraintreeIntegrationService();
-        }
     }
 
     async execute(payload: OrderRequestBody, options?: PaymentRequestOptions): Promise<void> {
@@ -233,17 +222,9 @@ export default class ApplePayPaymentStrategy implements PaymentStrategy {
         promise: ApplePayPromise,
     ) {
         const { token } = event.payment;
-
-        let deviceSessionId: string | undefined;
-
-        if (paymentMethod.initializationData?.gateway === ApplePayGatewayType.BRAINTREE) {
-            deviceSessionId = await this._getBraintreeDeviceData();
-        }
-
         const payment: Payment = {
             methodId: paymentMethod.id,
             paymentData: {
-                deviceSessionId,
                 formattedPayload: {
                     apple_pay_token: {
                         payment_data: token.paymentData,
@@ -266,30 +247,5 @@ export default class ApplePayPaymentStrategy implements PaymentStrategy {
                 new NotInitializedError(NotInitializedErrorType.PaymentNotInitialized),
             );
         }
-    }
-
-    private async _getBraintreeDeviceData() {
-        const data = await this._braintreeIntegrationService.getDataCollector();
-
-        return data.deviceData;
-    }
-
-    private async _initializeBraintreeIntegrationService() {
-        const state = await this._paymentIntegrationService.loadPaymentMethod(
-            ApplePayGatewayType.BRAINTREE,
-        );
-
-        const braintreePaymentMethod: PaymentMethod = state.getPaymentMethodOrThrow(
-            ApplePayGatewayType.BRAINTREE,
-        );
-
-        if (!braintreePaymentMethod.clientToken || !braintreePaymentMethod.initializationData) {
-            return;
-        }
-
-        this._braintreeIntegrationService.initialize(
-            braintreePaymentMethod.clientToken,
-            braintreePaymentMethod.initializationData,
-        );
     }
 }
