--- packages/apple-pay-integration/src/apple-pay.ts
@@ -1,3 +1,0 @@
-export enum ApplePayGatewayType {
-    BRAINTREE = 'braintree',
-}
