+++ packages/braintree-utils/src/mocks/braintree.mock.ts
@@ -1,5 +1,3 @@
-import { PaymentMethod } from '@bigcommerce/checkout-sdk/payment-integration-api';
-
 import {
     BraintreeClient,
     BraintreeConnect,
@@ -19,29 +17,6 @@ export function getClientMock(): BraintreeClient {
     return {
         request: jest.fn(),
         getVersion: jest.fn(),
-    };
-}
-
-export function getBraintree(): PaymentMethod {
-    return {
-        id: 'braintree',
-        clientToken: 'clientToken',
-        logoUrl:
-            'https://cdn.bcapp.dev/rHEAD/modules/checkout/braintree/images/paypal_powered_braintree_horizontal.png',
-        method: 'credit-card',
-        supportedCards: ['VISA', 'MC', 'AMEX', 'DISCOVER', 'JCB', 'DINERS'],
-        config: {
-            displayName: 'Credit Card',
-            cardCode: true,
-            enablePaypal: true,
-            merchantId: '',
-            testMode: true,
-            isVisaCheckoutEnabled: false,
-        },
-        initializationData: {
-            isAcceleratedCheckoutEnabled: false,
-        },
-        type: 'PAYMENT_TYPE_API',
     };
 }
